Fix image names for unquoted datafiles
String.substring will return a string of length end-start, from index 'start' to index 'end-1'. This means that for sprite "ship/image", the previous end-1 behavior was not needed to trim off the newline separator, and would actually strip off the closing ". However, if there is no closing ", this would strip off the last character of the name (e.g. sprite ship/archon -> ship/archo)

The `replace` operator is greedy, and should match both the starting and ending " to replace them.