Made changes to `logCommand` and James now posts a timeout message in 'the-corner`
The logCommand function now has a third argument, a string that contains the name of the room to post the message to. when an empty string is passed, it will default to `mod-log` but this is used to make James post a copy of the message sent to `mod-log` to `the-corner` so people can know how long they've been timed out for and (if one was given) the reason for it.
Alternatively, I could add some code to the timeout function to send the message to the corner instead of modifying and using the logCommand function.